pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

# Include ALL relevant directories and files
[tool.setuptools.packages.find]
include = [
    "qualitative_analysis*",
    "notebooks*",
    "codebook*",
    "data*"
]

# Include non-code files (e.g., CSVs, JSONs, notebooks)
[tool.setuptools.package-data]
qualitative_analysis = ["*.json", "*.csv", "*.txt"]
notebooks = ["*.ipynb"]
codebook = ["*"]
data = ["*"]

# Include app.py as a console script
[project.scripts]
qualitative-app = "app:main"

[project]
name = "qualitative_analysis"
version = "0.1.0"
description = "A toolkit for qualitative data analysis using language models."
readme = "README.md"
requires-python = ">=3.8"
license = { text = "UNLICENSED" }

authors = [
  { name="Olivier Clerc", email="oclerc38@gmail.com" },
]

dependencies = [
  "pandas==2.2.3",
  "scikit-learn==1.5.2",
  "seaborn==0.13.2",
  "streamlit==1.40.2",
  "together==1.3.5",
  "python-dotenv==1.0.1",
  "matplotlib==3.9.2",
  "chardet==5.2.0",
  "openai==1.109.1",
  "httpx==0.28.1",
  "regex==2024.9.11",
  "transformers==4.48.3",
  "torch==2.6.0",
  "torchvision==0.21.0",
  "sentencepiece==0.2.0",
  "sacremoses==0.1.1",
  "openpyxl==3.1.5",
  "krippendorff==0.8.1",
  "typing_extensions>=4.13.2",
  "google-genai==1.50.1",
  "xlsxwriter==3.2.0",
  "transformers==4.48.3",
  "openpyxl==3.1.5",
  "anthropic==0.49.0",
]

# [project.optional-dependencies]
# dev = [
#   "pytest",
#   "mypy",
#   "black",
#   "flake8",
# ]
